POI：bug 58775: use short for data format index, int for cell style index

+++ main/HSSF/Util/HSSFCellUtil.cs
@@ -369,9 +369,9 @@ namespace NPOI.HSSF.Util
 
             // index seems like what  index the cellstyle is in the list of styles for a workbook.
             // not good to compare on!
-            short numberCellStyles = workbook.NumCellStyles;
+            int numberCellStyles = workbook.NumCellStyles;
 
-            for (short i = 0; i < numberCellStyles; i++)
+            for (int i = 0; i < numberCellStyles; i++)
             {
                 ICellStyle wbStyle = workbook.GetCellStyleAt(i);
                 Hashtable wbStyleMap = GetFormatProperties(wbStyle);
